Some corrections in flags and many questions
Line 121: "You should read the documentation of each command to know what flags are available." I have googled for half an hour and cannot find general, easy-to-read documentation about the command line and using flags. It is specified for Go or for Google Chromium etc. etc., but where can they find a list with what the flags are and what they do? This is for Linux: https://websiteforstudents.com/use-the-rm-command-to-delete-files-and-folders-on-ubuntu-18-04-with-examples/ But can our students use this for more information?
"if you give him -f it gets powers and can delete no matter what, and if you give him -rf it, well, it can delete folders no matter what". I tried to find online what -f and -rf do but it is hard to understand the differences between rm -r and rm -f and rm -rf. We need to explain this better.