Take out screenshots with Linux commands
I have taken out the screen shots in the section 'More commands' and created a table there. I took out the word 'Linux' because we use these commands in the git bash and OS X as well. I clarified parts that were not easy to understand and took out information that is not relevant to beginners.