Update lesson0.md
I think we should add the VSC setting of autosave. So many times things go wrong because students forget to save their changes.

Line 320: #### Open your VS project in a Browser with one click
This item does not have any text so I took it out.

I have taken out  * There are GUI software - Gitkraken is one of them https://www.gitkraken.com/ because a) they will not know what GUI software is, and b) we have to keep them working on the command line for now

I added a bit more explanation to the sections on Git and GitHub, to make this accessible to beginners who never heard of version control. I still think we go too fast into the Git commands, but maybe we can take more time in class to work slowly towards this, with more examples etc.